+++ src/filedb/inner/idx.rs
@@ -184,24 +184,6 @@ impl IdxFile {
         let keys_len = idx_node.get_ref().keys_len();
         Ok(KeysCount::new(keys_len.try_into().unwrap()))
     }
-    #[cfg(feature = "node_dm32")]
-    #[inline]
-    pub fn read_node_keys_len(&mut self, offset: NodeOffset) -> Result<usize> {
-        let mut locked = RefCell::borrow_mut(&self.0);
-        locked.read_node_keys_len(offset)
-    }
-    #[cfg(feature = "node_dm32")]
-    #[inline]
-    pub fn read_node_keys_get(&mut self, offset: NodeOffset, idx: usize) -> Result<RecordOffset> {
-        let mut locked = RefCell::borrow_mut(&self.0);
-        locked.read_node_keys_get(offset, idx)
-    }
-    #[cfg(feature = "node_dm32")]
-    #[inline]
-    pub fn read_node_downs_get(&mut self, offset: NodeOffset, idx: usize) -> Result<NodeOffset> {
-        let mut locked = RefCell::borrow_mut(&self.0);
-        locked.read_node_downs_get(offset, idx)
-    }
 }
 
 // for debug
@@ -990,7 +972,7 @@ db_map.depth_of_node_tree(): 8
 */
 pub const NODE_SLOTS_MAX: u16 = 12;
 
-#[cfg(not(any(feature = "small_node_slots", feature = "node_dm32")))]
+#[cfg(not(feature = "small_node_slots"))]
 const NODE_SIZE_ARY: [u32; 8] = [
     16 * 2,
     16 * 3,
@@ -1000,18 +982,6 @@ const NODE_SIZE_ARY: [u32; 8] = [
     16 * 7,
     16 * 8,
     16 * 9,
-];
-
-#[cfg(feature = "node_dm32")]
-const NODE_SIZE_ARY: [u32; 8] = [
-    16 * 5,
-    16 * 6,
-    16 * 7,
-    16 * 8,
-    16 * 9,
-    16 * 10,
-    16 * 11,
-    16 * 12,
 ];
 
 /*
@@ -1127,11 +1097,7 @@ impl VarFileNodeCache {
             #[cfg(any(feature = "vf_u32u32", feature = "vf_u64u64"))]
             let encoded_len = 4;
             #[cfg(feature = "vf_vu64")]
-            #[cfg(not(feature = "node_dm32"))]
-            let encoded_len = 2;
-            #[cfg(feature = "vf_vu64")]
-            #[cfg(feature = "node_dm32")]
-            let encoded_len = 4;
+            let encoded_len = if buf_len < 128 { 1 } else { 2 };
             //let encoded_len = vu64::encoded_len(buf_len as u64);
             //
             // buggy: size operation for node size.
@@ -1182,7 +1148,8 @@ impl VarFileNodeCache {
                 self.0.seek_from_start(free_node_offset)?;
                 let node_size = self.0.read_node_size()?;
                 debug_assert!(
-                    (new_node_size.as_value() > 512 && node_size >= new_node_size)
+                    (new_node_size.as_value() > NODE_SIZE_ARY[NODE_SIZE_ARY.len() - 1]
+                        && node_size >= new_node_size)
                         || node_size == new_node_size,
                     "node_size: {} == new_node_size: {}",
                     node_size.as_value(),
@@ -1306,111 +1273,6 @@ impl VarFileNodeCache {
         let node_ = self.1.put(&mut self.0, node_, node_size, false)?;
         //
         Ok(node_)
-    }
-
-    #[cfg(feature = "node_dm32")]
-    fn read_node_keys_len(&mut self, offset: NodeOffset) -> Result<usize> {
-        debug_assert!(!offset.is_zero());
-        debug_assert!((offset.as_value() & 0x0F) == 0);
-        //
-        /*
-        if let Some(cached_node) = self.1.get(&offset) {
-            return Ok(cached_node.keys_len());
-        }
-        */
-        //
-        self.0.seek_from_start(offset)?;
-        let node_size = self.0.read_node_size()?;
-        debug_assert!(
-            !node_size.is_zero(),
-            "!node_size.is_zero(), offset: {}",
-            offset
-        );
-        let keys_count = self.0.read_keys_count()?;
-        debug_assert!(
-            keys_count.as_value() < NODE_SLOTS_MAX,
-            "keys_count: {} < NODE_SLOTS_MAX",
-            keys_count
-        );
-        let keys_count: usize = keys_count.into();
-        Ok(keys_count)
-    }
-    #[cfg(feature = "node_dm32")]
-    fn read_node_keys_get(&mut self, offset: NodeOffset, idx: usize) -> Result<RecordOffset> {
-        debug_assert!(!offset.is_zero());
-        debug_assert!((offset.as_value() & 0x0F) == 0);
-        //
-        /*
-        if let Some(cached_node) = self.1.get(&offset) {
-            return Ok(cached_node.keys_len());
-        }
-        */
-        //
-        self.0.seek_from_start(offset)?;
-        let node_size = self.0.read_node_size()?;
-        debug_assert!(
-            !node_size.is_zero(),
-            "!node_size.is_zero(), offset: {}",
-            offset
-        );
-        let keys_count = self.0.read_keys_count()?;
-        debug_assert!(
-            keys_count.as_value() < NODE_SLOTS_MAX,
-            "keys_count: {} < NODE_SLOTS_MAX",
-            keys_count
-        );
-        let keys_count: usize = keys_count.into();
-        debug_assert!(idx < keys_count);
-        if idx >= keys_count {
-            return Ok(RecordOffset::new(0));
-        }
-        self.0.seek_skip_size(NodeSize::new(4 * idx as u32))?;
-        let record_offset = self
-            .0
-            .read_record_offset()
-            .unwrap_or_else(|_| panic!("offset:{}, i:{}", offset, idx));
-        debug_assert!(!record_offset.is_zero());
-        Ok(record_offset)
-    }
-    #[cfg(feature = "node_dm32")]
-    fn read_node_downs_get(&mut self, offset: NodeOffset, idx: usize) -> Result<NodeOffset> {
-        debug_assert!(!offset.is_zero());
-        debug_assert!((offset.as_value() & 0x0F) == 0);
-        //
-        /*
-        if let Some(cached_node) = self.1.get(&offset) {
-            return Ok(cached_node.keys_len());
-        }
-        */
-        //
-        self.0.seek_from_start(offset)?;
-        let node_size = self.0.read_node_size()?;
-        debug_assert!(
-            !node_size.is_zero(),
-            "!node_size.is_zero(), offset: {}",
-            offset
-        );
-        let keys_count = self.0.read_keys_count()?;
-        debug_assert!(
-            keys_count.as_value() < NODE_SLOTS_MAX,
-            "keys_count: {} < NODE_SLOTS_MAX",
-            keys_count
-        );
-        let keys_count: usize = keys_count.into();
-        debug_assert!(idx < keys_count + 1);
-        if idx >= keys_count + 1 {
-            return Ok(NodeOffset::new(0));
-        }
-        //
-        self.0
-            .seek_skip_size(NodeSize::new(4 * keys_count as u32 + 8 * idx as u32))?;
-        let node_offset = self
-            .0
-            .read_node_offset()
-            .map(|a| NodeOffset::new(a.into()))
-            .unwrap_or_else(|_| panic!("offset:{}, i:{}", offset, idx));
-        debug_assert!((node_offset.as_value() & 0x0F) == 0);
-        Ok(node_offset)
     }
 }