+++ src/test_main5.rs
@@ -1,30 +1,32 @@
 use siamesedb::filedb::{FileBufSizeParam, FileDbMapString, FileDbParams};
 use siamesedb::DbXxx;
+use std::str::FromStr;
 
 fn main() -> Result<(), std::io::Error> {
     let db_name = "target/tmp/testA5.siamesedb";
     //
     let args: Vec<String> = std::env::args().collect();
     match args[1].as_str() {
-        "-g" => _test_gen(db_name)?,
-        "-c" => _test_gen_check(db_name)?,
+        "-c" => _test_create(db_name)?,
+        "-w" => _test_write(db_name)?,
+        "-r" => _test_read(db_name)?,
         _ => {
-            eprintln!("[usage] {} {{-g|-c}}", args[0]);
+            eprintln!("[usage] {} {{-c|-w|-r}}", args[0]);
         }
     }
     Ok(())
 }
 
-fn conv_to_kv_string(i: i64) -> (String, String) {
-    let bytes = i.to_le_bytes();
+fn conv_to_kv_string(ki: i64, vi: i64) -> (String, String) {
+    let bytes = ki.to_le_bytes();
+    let k = format!("{}.{}.{}.{}", bytes[0], bytes[1], bytes[2], bytes[3]);
     //let k = format!("key-{}.{}.{}", bytes[0], bytes[1], bytes[2]);
-    let k = format!("key-{}.{}.{}", bytes[0], bytes[1], bytes[2]).repeat(4);
-    //let v = format!("val-{}", i);
-    let v = String::new();
-    (k, v)
+    //let k = format!("key-{}.{}.{}", bytes[0], bytes[1], bytes[2]).repeat(4);
+    let v = format!("{}", vi);
+    (k.into(), v)
 }
 
-fn _test_gen(db_name: &str) -> Result<(), std::io::Error> {
+fn _test_create(db_name: &str) -> Result<(), std::io::Error> {
     let _ = std::fs::remove_dir_all(db_name);
     let db = siamesedb::open_file(db_name).unwrap();
     let mut db_map = db
@@ -47,18 +49,20 @@ fn _test_gen(db_name: &str) -> Result<(), std::io::Error> {
     //
     db_map.read_fill_buffer()?;
     //
+    let vi: i64 = 0;
+    //
     let mut kv_vec: Vec<(String, String)> = Vec::new();
-    let mut i: i64 = 0;
+    let mut ki: i64 = 0;
     loop {
-        i += 1;
-        if i > 2_000_000 {
+        ki += 1;
+        if ki > 2_000_000 {
             break;
         }
-        if i % 10_000 == 0 {
+        if ki % 10_000 == 0 {
             db_map.bulk_put_string(&kv_vec)?;
             kv_vec.clear();
         }
-        let (k, v) = conv_to_kv_string(i);
+        let (k, v) = conv_to_kv_string(ki, vi);
         kv_vec.push((k, v));
     }
     if !kv_vec.is_empty() {
@@ -66,7 +70,62 @@ fn _test_gen(db_name: &str) -> Result<(), std::io::Error> {
     }
     db_map.flush()
 }
-fn _test_gen_check(db_name: &str) -> Result<(), std::io::Error> {
+
+fn _test_write(db_name: &str) -> Result<(), std::io::Error> {
+    let db = siamesedb::open_file(db_name).unwrap();
+    let mut db_map = db
+        .db_map_string_with_params(
+            "some_map1",
+            FileDbParams {
+                /*
+                dat_buf_size: FileBufSizeParam::PerMille(100),
+                idx_buf_size: FileBufSizeParam::PerMille(300),
+                */
+                dat_buf_size: FileBufSizeParam::PerMille(1000),
+                idx_buf_size: FileBufSizeParam::PerMille(1000),
+                /*
+                dat_buf_size: FileBufSizeParam::Auto,
+                idx_buf_size: FileBufSizeParam::Auto,
+                */
+            },
+        )
+        .unwrap();
+    //
+    db_map.read_fill_buffer()?;
+    //
+    let (k, _v) = conv_to_kv_string(1, 0);
+    let vi: i64 = {
+        if let Some(s) = db_map.get_string(&k)? {
+            match i64::from_str(&s) {
+                Ok(i) => i + 1,
+                Err(_) => 0,
+            }
+        } else {
+            0
+        }
+    };
+    //
+    let mut kv_vec: Vec<(String, String)> = Vec::new();
+    let mut ki: i64 = 0;
+    loop {
+        ki += 1;
+        if ki > 2_000_000 {
+            break;
+        }
+        if ki % 10_000 == 0 {
+            db_map.bulk_put_string(&kv_vec)?;
+            kv_vec.clear();
+        }
+        let (k, v) = conv_to_kv_string(ki, vi);
+        kv_vec.push((k, v));
+    }
+    if !kv_vec.is_empty() {
+        db_map.bulk_put_string(&kv_vec)?;
+    }
+    db_map.flush()
+}
+
+fn _test_read(db_name: &str) -> Result<(), std::io::Error> {
     let db = siamesedb::open_file(db_name).unwrap();
     let mut db_map = db
         .db_map_string_with_params(
@@ -88,26 +147,38 @@ fn _test_gen_check(db_name: &str) -> Result<(), std::io::Error> {
     //
     db_map.read_fill_buffer()?;
     //
+    let (k, _v) = conv_to_kv_string(1, 0);
+    let vi: i64 = {
+        if let Some(s) = db_map.get_string(&k)? {
+            match i64::from_str(&s) {
+                Ok(i) => i,
+                Err(_) => 0,
+            }
+        } else {
+            0
+        }
+    };
+    //
     let mut key_vec: Vec<String> = Vec::new();
     let mut value_vec: Vec<String> = Vec::new();
-    let mut i: i64 = 0;
+    let mut ki: i64 = 0;
     loop {
-        i += 1;
-        if i > 2_000_000 {
+        ki += 1;
+        if ki > 2_000_000 {
             break;
         }
-        if i % 10_000 == 0 {
-            _test_gen_check_one(&mut db_map, &key_vec, &value_vec)?;
+        if ki % 10_000 == 0 {
+            _test_read_one(&mut db_map, &key_vec, &value_vec)?;
             //
             key_vec.clear();
             value_vec.clear();
         }
-        let (k, correct) = conv_to_kv_string(i);
+        let (k, correct) = conv_to_kv_string(ki, vi);
         key_vec.push(k);
         value_vec.push(correct);
     }
     if !key_vec.is_empty() {
-        _test_gen_check_one(&mut db_map, &key_vec, &value_vec)?;
+        _test_read_one(&mut db_map, &key_vec, &value_vec)?;
         //
         key_vec.clear();
         value_vec.clear();
@@ -115,7 +186,7 @@ fn _test_gen_check(db_name: &str) -> Result<(), std::io::Error> {
     Ok(())
 }
 
-fn _test_gen_check_one(
+fn _test_read_one(
     db_map: &mut FileDbMapString,
     key_vec: &[String],
     value_vec: &[String],